wip implemented lin obj / next is (smoothed) const
need to figure out a reasonable way to implement the (smoothed)
constraints. I think that the smoothe constraint should be part of an
object which wraps over the various inputs of the constraint.

But then, should there be a difference between equality and inequality?
In practice, they are actually very similar.

Should they all be SeparableConstraint? Need to make a list of them and
see what they have in commmon.